elfmalloc: Moved frontends to separate module
