+++ elfmalloc/src/slag.rs
@@ -56,15 +56,17 @@
 //! placed in a global cache of dirty pages (mentioned above). From there it can be uncommitted or
 //! cached for use by other object sizes.
 //!
+//! See the `frontends` module for how the slag subsystem is used to construct allocators.
+//!
 //! [1]: https://arxiv.org/abs/1503.09006
 use std::mem;
 use std::sync::atomic::{fence, AtomicPtr, AtomicUsize, Ordering};
 use super::bagpipe::bag::{Revocable, WeakBag};
 use super::bagpipe::{BagPipe, BagCleanup};
 use super::bagpipe::queue::{FAAQueueLowLevel, RevocableFAAQueue};
-use super::utils::{mmap, LazyInitializable, OwnedArray, likely, unlikely};
+use super::utils::{mmap, LazyInitializable, unlikely};
 use super::alloc_type::AllocType;
-use super::sources::{MemorySource, MmapSource};
+use super::sources::MemorySource;
 use std::marker::PhantomData;
 use std::ptr;
 use std::cmp;
@@ -119,46 +121,295 @@ where
     fn backing_memory(&self) -> &Self::Block;
 }
 
-/// Metadata about a particular size-class of objects allocated to a particular page size.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct Metadata {
-    /// The size of an individual object (in bytes).
-    pub object_size: usize,
 
-    /// Number of objects in a slab.
-    n_objects: usize,
-    /// Number of `usize` values comprising the bit-set.
-    n_bitset_words: usize,
-    /// Total bytes for the slag.
-    total_bytes: usize,
-    /// Offset (in bytes) from the start of the slag to get the start of the bit-set.
-    bitset_offset: isize,
-    /// Offset (in bytes) from the start of the slag to get to the first object.
-    objects_offset: isize,
-    /// The mask used to initialize a `Slag` bit-set.
-    object_mask: usize,
-    /// The base-2 log of the amount of memory represented by a bit in a bit-set.
-    bit_rep_shift: usize,
-    /// Metadata objects often represent one of several available object sizes. These allocators
-    /// are used in concert to provide a general dynamic memory allocator. Threads store pointers
-    /// to each of these allocators in an array, this is the index into that array.
-    local_index: usize,
+pub use self::metadata::{Metadata, compute_metadata};
 
-    /// A per-allocator tunable that indicates how many available objects a slag must have before
-    /// it can be made available to allocating threads.
-    cutoff_objects: usize,
+mod metadata {
+    use super::*;
+    /// Metadata about a particular size-class of objects allocated to a particular page size.
+    #[derive(Copy, Clone, Debug, Eq, PartialEq)]
+    pub struct Metadata {
+        /// The size of an individual object (in bytes).
+        pub object_size: usize,
 
-    /// Size that can be used by this particular object class.
+        /// Number of objects in a slab.
+        pub n_objects: usize,
+        /// Number of `usize` values comprising the bit-set.
+        pub n_bitset_words: usize,
+        /// Total bytes for the slag.
+        pub total_bytes: usize,
+        /// Offset (in bytes) from the start of the slag to get the start of the bit-set.
+        pub bitset_offset: isize,
+        /// Offset (in bytes) from the start of the slag to get to the first object.
+        pub objects_offset: isize,
+        /// The mask used to initialize a `Slag` bit-set.
+        pub object_mask: usize,
+        /// The base-2 log of the amount of memory represented by a bit in a bit-set.
+        pub bit_rep_shift: usize,
+        /// Metadata objects often represent one of several available object sizes. These allocators
+        /// are used in concert to provide a general dynamic memory allocator. Threads store pointers
+        /// to each of these allocators in an array, this is the index into that array.
+        local_index: usize,
+
+        /// A per-allocator tunable that indicates how many available objects a slag must have before
+        /// it can be made available to allocating threads.
+        pub cutoff_objects: usize,
+
+        /// Size that can be used by this particular object class.
+        ///
+        /// It may be advantageous to store smaller size classes in a large contiguous unit of memory,
+        /// but leave a large portion of that memory unused (with the expectation that it is
+        /// uncommited).
+        pub usable_size: usize,
+
+        pub ty: AllocType,
+    }
+
+    /// Compute an optimal layout for objects of size `obj_size` for `Slag`s of size `page_size` with
+    /// cutoff a `cutoff_factor` fraction of total objects, and a local index `local_index`.
     ///
-    /// It may be advantageous to store smaller size classes in a large contiguous unit of memory,
-    /// but leave a large portion of that memory unused (with the expectation that it is
-    /// uncommited).
-    usable_size: usize,
+    /// This function essentially performs an exhaustive search over the possible number of objects and
+    /// the possible values of `bit_rep_shift` and picks the one with the lowest ratio of unused space,
+    /// with a tie-braking preference for fewer words in the `Slag` bit-set.
+    pub fn compute_metadata(
+        obj_size: usize,
+        page_size: usize,
+        local_index: usize,
+        cutoff_factor: f64,
+        usable_size: usize,
+        ty: AllocType,
+    ) -> Metadata {
+        // This is by far the ugliest function in this project. It is all plumbing, heuristics, and
+        // other gross things.
+        // We start with a bunch of useful helper functions:
 
-    ty: AllocType,
+        /// Calculate the number of bytes in the bitset needed to represent `n_objects` objects, using
+        /// `gran` bits per object.
+        fn bitset_bytes(n_objects: usize, gran: usize) -> usize {
+            let word_size = mem::size_of::<Word>();
+            let word_bits = Word::bits();
+
+            let bits = n_objects * gran;
+            let words = if bits % word_bits == 0 {
+                bits / word_bits
+            } else {
+                bits / word_bits + 1
+            };
+            words * word_size
+        }
+
+        /// Calculate the padding (in bytes) required to ensure the `Slag` objects are aligned to
+        /// `alignment` bytes, given `n_objects` objects and `gran` bits per object in the bit-set.
+        fn align_padding(alignment: usize, n_objects: usize, gran: usize) -> usize {
+            debug_assert!(alignment.is_power_of_two());
+            let header_size = slag_size();
+            let h_bitset_size = header_size + bitset_bytes(n_objects, gran);
+            let rounded = (h_bitset_size + (alignment - 1)) & !(alignment - 1);
+            rounded - h_bitset_size
+        }
+
+        /// Compute the total bytes used for `n_objects` objects each of size `size` bytes represented
+        /// by `gran` bits in the bit-set. This function includes the heuristic that all power-of-two
+        /// sizes are aligned to their size, inserting padding accordingly.
+        fn total_bytes(size: usize, gran: usize, n_objects: usize) -> usize {
+            let header_size = slag_size();
+            let padding = if size.is_power_of_two() {
+                align_padding(size, n_objects, gran)
+            } else {
+                0
+            };
+            header_size + bitset_bytes(n_objects, gran) + padding + n_objects * size
+        }
+
+        /// Perform an exhaustive search for the lowest-fragmentation layout of objects of a particular
+        /// size, and "shadow size" (i.e. `1 << round_up_to_shift`).
+        ///
+        /// `cutoff_factor` and `local_index` are just passing through configuration parameters.
+        /// `usable_size` is used for smaller object sizes in order for them to use a smaller amount of
+        /// memory than the total page size.
+        ///
+        /// TODO(ezrosent): using a builder for Metadata would clean things up considerably.
+        fn meta_inner(
+            size: usize,
+            page_size: usize,
+            round_up_to_shift: usize,
+            local_index: usize,
+            cutoff_factor: f64,
+            usable_size: usize,
+        ) -> (f64, usize, Metadata) {
+            use std::cmp;
+            let usable_size = cmp::min(usable_size, page_size);
+            let mut mult = 1.0;
+            // we round up to the nearest round_up_to_bytes. As a result, there may be a small amount
+            // of padding per-object in the `Slag`.
+            let round_up_to_bytes = 1 << round_up_to_shift;
+            let padding_per_object = {
+                let rem = size % round_up_to_bytes;
+                if rem == 0 { 0 } else { round_up_to_bytes - rem }
+            };
+            let padded_size = size + padding_per_object;
+            // gran (read "granularity") is the number of bits used to represent a single object in the
+            // slag bit-set.
+            let gran = padded_size / round_up_to_bytes;
+            debug_assert!(usable_size > 0);
+            debug_assert!(round_up_to_bytes > 0);
+            debug_assert!(round_up_to_bytes.is_power_of_two());
+            debug_assert!(gran > 0);
+            #[cfg_attr(feature = "cargo-clippy", allow(panic_params))]
+            debug_assert!({
+                if gran == 1 {
+                    padded_size.is_power_of_two()
+                } else {
+                    true
+                }
+            });
+            // == gran
+            // TODO(ezrosent): remove one of these
+            let bits_per_object = padded_size >> round_up_to_shift;
+            let bits_per_word = Word::bits();
+            // Let's say we are storing 24-byte objects with round_up_to_shift=3. This results in each
+            // bit in the bit-set representing 8-byte chunks, with each object being 3 bits in the
+            // bitset.
+            //
+            // On 64-bit machines, this means that objects do not fit exactly into a single word: after
+            // 21 objects we will have used 63 bits, meaning the 22nd object uses the last bit but is
+            // using 2 extra bits in the next bitset word. This number '2' is referred to as "slush".
+            //
+            // During initialization we perform a word_bits + slush_size (in this case, 66)-bit
+            // rotation to compute the masks used for full `Slags`. Part of this is holding the extra
+            // "slush" bits in their own mask. This mask is a `usize`, so slush_size cannot exceed
+            // word_bits.
+            //
+            // NB: Having a slush_size greater than the number of bits in a word is a degenerate case,
+            // as it implies that more than word_bits bits are required to represent a single object in
+            // a bitset. This is essentially never a good idea.
+            let slush_size = bits_per_object - (bits_per_word % bits_per_object);
+            if bits_per_word < slush_size {
+                // give configurations violating this invariant a negative score
+                mult = -1.0;
+            }
+            // First, find out how many objects we can fit while using a maximum of `usable_size`
+            let mut n_objects = 1;
+            loop {
+                if total_bytes(padded_size, gran, n_objects + 1) > usable_size {
+                    break;
+                }
+                n_objects += 1;
+            }
+            // Get the alignment padding we are using. Note that this is already computed in
+            // `total_bytes`, we are just extracting it here.
+            let align_padding = if padded_size.is_power_of_two() {
+                align_padding(padded_size, n_objects, gran)
+            } else {
+                0
+            };
+
+            // This is takes all of the space we use in this configuration and subtracts all of
+            // the "cruft" that isn't used to actually store an object.
+            let bs = (total_bytes(padded_size, gran, n_objects) - n_objects * padding_per_object -
+                          bitset_bytes(n_objects, gran) - slag_size() -
+                          align_padding) as f64;
+            let score = if bs > usable_size as f64 { -1.0 } else { 1.0 } * bs /
+                (usable_size as f64);
+            let header_offset = slag_size() as isize;
+            let n_words = bitset_bytes(n_objects, gran) / mem::size_of::<Word>();
+            (
+                score * mult,
+                n_words,
+                Metadata {
+                    n_objects: n_objects,
+                    n_bitset_words: n_words,
+                    total_bytes: page_size,
+                    bitset_offset: header_offset,
+                    objects_offset: header_offset +
+                        (align_padding + bitset_bytes(n_objects, gran)) as isize,
+                    object_size: padded_size,
+                    object_mask: 1,
+                    bit_rep_shift: round_up_to_bytes.trailing_zeros() as usize,
+                    local_index: local_index,
+                    cutoff_objects: cmp::max(1, (n_objects as f64 * cutoff_factor) as usize),
+                    usable_size: usable_size,
+                    ty: AllocType::SmallSlag,
+                },
+            )
+        }
+        let test_meta = Metadata {
+            n_objects: 0,
+            n_bitset_words: 0,
+            total_bytes: 0,
+            bitset_offset: 0,
+            objects_offset: 0,
+            object_size: 0,
+            object_mask: 0,
+            bit_rep_shift: 0,
+            local_index: 0,
+            cutoff_objects: 0,
+            usable_size: 0,
+            ty: AllocType::SmallSlag,
+        };
+
+        // now we perform an exhaustive search over these elements.
+
+        // unused in release builds
+        #[allow(unused)]
+        let (frag, _, mut meta) = (1..(obj_size.next_power_of_two().trailing_zeros() as usize + 1))
+            .map(|shift| {
+                meta_inner(
+                    obj_size,
+                    page_size,
+                    shift,
+                    local_index,
+                    cutoff_factor,
+                    usable_size,
+                )
+            })
+            .fold((-10.0, 1000, test_meta), |o1, o2| if o1.0 < o2.0 ||
+                (o1.0 - o2.0).abs() < 1e-5 &&
+                    o1.1 > o2.1
+            {
+                o2
+            } else {
+                o1
+            });
+        // Compute the mask used to represent the first bitset word
+        let bits = Word::bits();
+        let bits_per_object = meta.object_size >> meta.bit_rep_shift;
+        let mut cur_bit = 0;
+        let mut mask = 0;
+        while cur_bit < bits {
+            mask |= 1 << cur_bit;
+            cur_bit += bits_per_object;
+        }
+        meta.object_mask = mask;
+        meta.ty = ty;
+        trace!("created {:?} fragmentation: {:?}", meta, frag);
+
+        meta
+    }
+
+    #[cfg(test)]
+    mod tests {
+        extern crate env_logger;
+        use super::*;
+        #[test]
+        fn metadata_basic() {
+            let _ = env_logger::init();
+
+            compute_metadata(8, 4096, 0, 0.8, 4, AllocType::SmallSlag);
+            compute_metadata(16, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
+            compute_metadata(24, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
+            compute_metadata(127, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
+            compute_metadata(800, 2 << 20, 0, 0.8, 32 << 10, AllocType::SmallSlag);
+            compute_metadata(514, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
+            compute_metadata(513, 2 << 20, 0, 0.8, 1024, AllocType::SmallSlag);
+            compute_metadata(768, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
+            compute_metadata(800, 4096, 0, 0.8, 32 << 10, AllocType::SmallSlag);
+            compute_metadata(1025, 4096, 0, 0.8, 32 << 10, AllocType::SmallSlag);
+        }
+    }
 }
 
-use self::bitset::Word;
+pub use self::bitset::Word;
 mod bitset {
     use std::sync::atomic::AtomicUsize;
     use std::mem;
@@ -182,7 +433,7 @@ mod bitset {
     }
 }
 
-use self::ref_count::RefCount;
+pub use self::ref_count::RefCount;
 mod ref_count {
     //! Implementation of the `Slag` reference count. This is a
     //! specialized wrapper around an `AtomicUsize` that also reserves a
@@ -292,7 +543,7 @@ pub struct Slag {
     /// structure of the bitset.
     ty: AllocType,
     meta: AtomicPtr<Metadata>,
-    rc: RefCount,
+    pub rc: RefCount,
     // for BagPipe revocation.
     handle: AtomicUsize,
 }
@@ -308,226 +559,7 @@ impl Revocable for Slag {
     }
 }
 
-/// Compute an optimal layout for objects of size `obj_size` for `Slag`s of size `page_size` with
-/// cutoff a `cutoff_factor` fraction of total objects, and a local index `local_index`.
-///
-/// This function essentially performs an exhaustive search over the possible number of objects and
-/// the possible values of `bit_rep_shift` and picks the one with the lowest ratio of unused space,
-/// with a tie-braking preference for fewer words in the `Slag` bit-set.
-pub fn compute_metadata(
-    obj_size: usize,
-    page_size: usize,
-    local_index: usize,
-    cutoff_factor: f64,
-    usable_size: usize,
-    ty: AllocType,
-) -> Metadata {
-    // This is by far the ugliest function in this project. It is all plumbing, heuristics, and
-    // other gross things.
-    // We start with a bunch of useful helper functions:
 
-    /// Calculate the number of bytes in the bitset needed to represent `n_objects` objects, using
-    /// `gran` bits per object.
-    fn bitset_bytes(n_objects: usize, gran: usize) -> usize {
-        let word_size = mem::size_of::<Word>();
-        let word_bits = Word::bits();
-
-        let bits = n_objects * gran;
-        let words = if bits % word_bits == 0 {
-            bits / word_bits
-        } else {
-            bits / word_bits + 1
-        };
-        words * word_size
-    }
-
-    /// Calculate the padding (in bytes) required to ensure the `Slag` objects are aligned to
-    /// `alignment` bytes, given `n_objects` objects and `gran` bits per object in the bit-set.
-    fn align_padding(alignment: usize, n_objects: usize, gran: usize) -> usize {
-        debug_assert!(alignment.is_power_of_two());
-        let header_size = slag_size();
-        let h_bitset_size = header_size + bitset_bytes(n_objects, gran);
-        let rounded = (h_bitset_size + (alignment - 1)) & !(alignment - 1);
-        rounded - h_bitset_size
-    }
-
-    /// Compute the total bytes used for `n_objects` objects each of size `size` bytes represented
-    /// by `gran` bits in the bit-set. This function includes the heuristic that all power-of-two
-    /// sizes are aligned to their size, inserting padding accordingly.
-    fn total_bytes(size: usize, gran: usize, n_objects: usize) -> usize {
-        let header_size = slag_size();
-        let padding = if size.is_power_of_two() {
-            align_padding(size, n_objects, gran)
-        } else {
-            0
-        };
-        header_size + bitset_bytes(n_objects, gran) + padding + n_objects * size
-    }
-
-    /// Perform an exhaustive search for the lowest-fragmentation layout of objects of a particular
-    /// size, and "shadow size" (i.e. `1 << round_up_to_shift`).
-    ///
-    /// `cutoff_factor` and `local_index` are just passing through configuration parameters.
-    /// `usable_size` is used for smaller object sizes in order for them to use a smaller amount of
-    /// memory than the total page size.
-    ///
-    /// TODO(ezrosent): using a builder for Metadata would clean things up considerably.
-    fn meta_inner(
-        size: usize,
-        page_size: usize,
-        round_up_to_shift: usize,
-        local_index: usize,
-        cutoff_factor: f64,
-        usable_size: usize,
-    ) -> (f64, usize, Metadata) {
-        use std::cmp;
-        let usable_size = cmp::min(usable_size, page_size);
-        let mut mult = 1.0;
-        // we round up to the nearest round_up_to_bytes. As a result, there may be a small amount
-        // of padding per-object in the `Slag`.
-        let round_up_to_bytes = 1 << round_up_to_shift;
-        let padding_per_object = {
-            let rem = size % round_up_to_bytes;
-            if rem == 0 { 0 } else { round_up_to_bytes - rem }
-        };
-        let padded_size = size + padding_per_object;
-        // gran (read "granularity") is the number of bits used to represent a single object in the
-        // slag bit-set.
-        let gran = padded_size / round_up_to_bytes;
-        debug_assert!(usable_size > 0);
-        debug_assert!(round_up_to_bytes > 0);
-        debug_assert!(round_up_to_bytes.is_power_of_two());
-        debug_assert!(gran > 0);
-        #[cfg_attr(feature = "cargo-clippy", allow(panic_params))]
-        debug_assert!({
-            if gran == 1 {
-                padded_size.is_power_of_two()
-            } else {
-                true
-            }
-        });
-        // == gran
-        // TODO(ezrosent): remove one of these
-        let bits_per_object = padded_size >> round_up_to_shift;
-        let bits_per_word = Word::bits();
-        // Let's say we are storing 24-byte objects with round_up_to_shift=3. This results in each
-        // bit in the bit-set representing 8-byte chunks, with each object being 3 bits in the
-        // bitset.
-        //
-        // On 64-bit machines, this means that objects do not fit exactly into a single word: after
-        // 21 objects we will have used 63 bits, meaning the 22nd object uses the last bit but is
-        // using 2 extra bits in the next bitset word. This number '2' is referred to as "slush".
-        //
-        // During initialization we perform a word_bits + slush_size (in this case, 66)-bit
-        // rotation to compute the masks used for full `Slags`. Part of this is holding the extra
-        // "slush" bits in their own mask. This mask is a `usize`, so slush_size cannot exceed
-        // word_bits.
-        //
-        // NB: Having a slush_size greater than the number of bits in a word is a degenerate case,
-        // as it implies that more than word_bits bits are required to represent a single object in
-        // a bitset. This is essentially never a good idea.
-        let slush_size = bits_per_object - (bits_per_word % bits_per_object);
-        if bits_per_word < slush_size {
-            // give configurations violating this invariant a negative score
-            mult = -1.0;
-        }
-        // First, find out how many objects we can fit while using a maximum of `usable_size`
-        let mut n_objects = 1;
-        loop {
-            if total_bytes(padded_size, gran, n_objects + 1) > usable_size {
-                break;
-            }
-            n_objects += 1;
-        }
-        // Get the alignment padding we are using. Note that this is already computed in
-        // `total_bytes`, we are just extracting it here.
-        let align_padding = if padded_size.is_power_of_two() {
-            align_padding(padded_size, n_objects, gran)
-        } else {
-            0
-        };
-
-        // This is takes all of the space we use in this configuration and subtracts all of
-        // the "cruft" that isn't used to actually store an object.
-        let bs = (total_bytes(padded_size, gran, n_objects) - n_objects * padding_per_object -
-                      bitset_bytes(n_objects, gran) - slag_size() -
-                      align_padding) as f64;
-        let score = if bs > usable_size as f64 { -1.0 } else { 1.0 } * bs / (usable_size as f64);
-        let header_offset = slag_size() as isize;
-        let n_words = bitset_bytes(n_objects, gran) / mem::size_of::<Word>();
-        (
-            score * mult,
-            n_words,
-            Metadata {
-                n_objects: n_objects,
-                n_bitset_words: n_words,
-                total_bytes: page_size,
-                bitset_offset: header_offset,
-                objects_offset: header_offset +
-                    (align_padding + bitset_bytes(n_objects, gran)) as isize,
-                object_size: padded_size,
-                object_mask: 1,
-                bit_rep_shift: round_up_to_bytes.trailing_zeros() as usize,
-                local_index: local_index,
-                cutoff_objects: cmp::max(1, (n_objects as f64 * cutoff_factor) as usize),
-                usable_size: usable_size,
-                ty: AllocType::SmallSlag,
-            },
-        )
-    }
-    let test_meta = Metadata {
-        n_objects: 0,
-        n_bitset_words: 0,
-        total_bytes: 0,
-        bitset_offset: 0,
-        objects_offset: 0,
-        object_size: 0,
-        object_mask: 0,
-        bit_rep_shift: 0,
-        local_index: 0,
-        cutoff_objects: 0,
-        usable_size: 0,
-        ty: AllocType::SmallSlag,
-    };
-
-    // now we perform an exhaustive search over these elements.
-
-    // unused in release builds
-    #[allow(unused)]
-    let (frag, _, mut meta) = (1..(obj_size.next_power_of_two().trailing_zeros() as usize + 1))
-        .map(|shift| {
-            meta_inner(
-                obj_size,
-                page_size,
-                shift,
-                local_index,
-                cutoff_factor,
-                usable_size,
-            )
-        })
-        .fold((-10.0, 1000, test_meta), |o1, o2| if o1.0 < o2.0 ||
-            (o1.0 - o2.0).abs() < 1e-5 &&
-                o1.1 > o2.1
-        {
-            o2
-        } else {
-            o1
-        });
-    // Compute the mask used to represent the first bitset word
-    let bits = Word::bits();
-    let bits_per_object = meta.object_size >> meta.bit_rep_shift;
-    let mut cur_bit = 0;
-    let mut mask = 0;
-    while cur_bit < bits {
-        mask |= 1 << cur_bit;
-        cur_bit += bits_per_object;
-    }
-    meta.object_mask = mask;
-    meta.ty = ty;
-    trace!("created {:?} fragmentation: {:?}", meta, frag);
-
-    meta
-}
 
 /// Given an index into a bitset, return the word of the bitset it is in, as well as the bit-wise
 /// index into that word to which the item corresponds.
@@ -548,11 +580,11 @@ fn split_index(item_index: usize) -> (isize, usize) {
 }
 
 /// An iterator over a claimed `Slag`'s bitset.
-struct AllocIter {
+pub struct AllocIter {
     /// The current word of the bitset being allocated from.
-    cur_word: usize,
+    pub cur_word: usize,
     /// A pointer to the next word that will be consumed, or one-past-the-end of the final word.
-    next_word: *mut Word,
+    pub next_word: *mut Word,
     /// A pointer to the corresponding `Slag`'s `RefCount`. This allows the iterator to decrement
     /// the `RefCount` when it consumes a new word.
     refcnt: *const RefCount,
@@ -598,17 +630,15 @@ impl AllocIter {
     /// The key step here is to perform an atomic fetch-and for the value 0. This will give
     /// us a snapshot of available objects corresponding to this word without the risk of losing
     /// any that are concurrently being freed.
-    fn refresh_word(&mut self) {
-        unsafe {
-            let next = self.next_word.as_ref().expect(
-                "bitset must point to valid memory",
-            );
-            self.next_word = self.next_word.offset(1);
-            self.cur_word = next.swap(0, Ordering::Acquire);
-            (*self.refcnt).dec_n(self.cur_word.count_ones() as usize);
-            self.remaining_words -= 1;
-            self.cur_word_index += 1;
-        }
+    pub unsafe fn refresh_word(&mut self) {
+        let next = self.next_word.as_ref().expect(
+            "bitset must point to valid memory",
+        );
+        self.next_word = self.next_word.offset(1);
+        self.cur_word = next.swap(0, Ordering::Acquire);
+        (*self.refcnt).dec_n(self.cur_word.count_ones() as usize);
+        self.remaining_words -= 1;
+        self.cur_word_index += 1;
     }
 }
 
@@ -619,13 +649,15 @@ impl Iterator for AllocIter {
         let word_size = Word::bits();
         loop {
             let next_bit = self.cur_word.trailing_zeros() as usize;
-            if unsafe { unlikely(next_bit == word_size) } {
-                if self.remaining_words == 0 {
-                    return None;
+            unsafe {
+                if unlikely(next_bit == word_size) {
+                    if self.remaining_words == 0 {
+                        return None;
+                    }
+                    self.refresh_word();
+                    continue;
                 }
-                self.refresh_word();
-                continue;
-            }
+            };
             unsafe {
                 self.cur_word ^= 1 << next_bit;
                 let object = self.object_base.offset(
@@ -638,25 +670,24 @@ impl Iterator for AllocIter {
     }
 }
 
-enum Transition {
+pub enum Transition {
     Null,
     Available,
     Full,
 }
 
 macro_rules! or_slag_word {
-        ($slag:expr, $bitset_offset:expr, $word:expr, $mask:expr) => {
-            {
-                let word = $word;
-                (($slag as *mut u8).offset($bitset_offset) as *mut Word)
-                    // go to the bitset we want
-                    .offset(word as isize)
-                    .as_ref()
-                    .expect("slag word null")
-                    .store($mask, Ordering::Relaxed);
-            }
-
-        };
+    ($slag:expr, $bitset_offset:expr, $word:expr, $mask:expr) => {
+        {
+            let word = $word;
+            (($slag as *mut u8).offset($bitset_offset) as *mut Word)
+                // go to the bitset we want
+                .offset(word as isize)
+                .as_ref()
+                .expect("slag word null")
+                .store($mask, Ordering::Relaxed);
+        }
+    };
 }
 
 impl Slag {
@@ -672,7 +703,7 @@ impl Slag {
         }
     }
 
-    fn as_raw(&self) -> *mut Self {
+    pub fn as_raw(&self) -> *mut Self {
         self as *const _ as *mut Self
     }
 
@@ -682,7 +713,7 @@ impl Slag {
     /// already initialized has been `claim`ed. As such, this simply amounts to initializing all of
     /// the `Slag` data-structures. In order to work in complete generality, the bit-set
     /// initialization is a bit subtle.
-    unsafe fn init(slag: *mut Self, meta: &Metadata) {
+    pub unsafe fn init(slag: *mut Self, meta: &Metadata) {
         let slf = slag.as_mut().expect("null slag");
         slf.set_metadata(meta as *const _ as *mut Metadata);
         ptr::write(&mut slf.ty, meta.ty);
@@ -765,7 +796,7 @@ impl Slag {
     }
 
     #[inline]
-    fn get_word(raw_self: *mut Slag, item: *mut u8, m: &Metadata) -> (isize, usize) {
+    pub fn get_word(raw_self: *mut Slag, item: *mut u8, m: &Metadata) -> (isize, usize) {
         let it_num = item as usize;
         let self_num = raw_self as usize;
         // `Slag` bitsets operate by "pretending" objects are not of the actual object size, but are
@@ -784,7 +815,7 @@ impl Slag {
     ///
     /// This method assumes `item` is a member of `self` (enforced in debug builds). It also
     /// computes whether or not this `free` operation triggered a state transition.
-    fn free(&self, item: *mut u8) -> Transition {
+    pub fn free(&self, item: *mut u8) -> Transition {
         let m = self.get_metadata();
         // must be in-bounds
         debug_assert!((item as usize) < (self.as_raw() as usize + m.total_bytes));
@@ -820,7 +851,7 @@ impl Slag {
     }
 
     /// Initialize an `AllocIter` for allocating out of the `Slag`.
-    fn refresh(&self, meta: &Metadata) -> AllocIter {
+    pub fn refresh(&self, meta: &Metadata) -> AllocIter {
         // offset calls are valid because size_of(u8) is 1
         unsafe {
             AllocIter::new(
@@ -834,297 +865,8 @@ impl Slag {
     }
 }
 
-/// A set data-structure used to batch remote free operations.
-struct Coalescer(OwnedArray<RemoteFreeCell>, PtrStack);
 
-/// The internal data for a `Coalescer`.
-///
-/// This holds all the provenance necessary to perform a batched remote free operation.
-struct RemoteFreeCell {
-    // TODO: explore using slag-specific information as well as word-specific.
-    // Idea: have the "key" be rc, and then use a *mut Slag to get at the entire bitset, along with
-    // an array of masks, then you can go as before, but only increment ref count once.
-    /// The reference count to be updated in accordance for the accumulated `mask`.
-    rc: *mut RefCount,
-    /// The specific bitset word to be or-ed with `mask`.
-    word: *mut Word,
-    /// The accumulated mask to update `word`.
-    mask: usize,
-}
 
-impl Default for RemoteFreeCell {
-    fn default() -> Self {
-        RemoteFreeCell {
-            rc: ptr::null_mut(),
-            word: ptr::null_mut(),
-            mask: 0,
-        }
-    }
-}
-
-impl Coalescer {
-    fn new(size: usize) -> Self {
-        Coalescer(
-            OwnedArray::new(size.next_power_of_two()),
-            PtrStack::new(size),
-        )
-    }
-
-    fn bucket_num(&self, word: usize) -> usize {
-        // we can do a "fast mod" operation because we know len() is a power of two (see `new`)
-        word & (self.0.len() - 1)
-    }
-
-    /// Try to insert `item`.
-    ///
-    /// The return value indicates if the value was successfully inserted.
-    unsafe fn insert(&mut self, item: *mut u8, meta: &Metadata) -> bool {
-        fn hash_ptr(p: *mut Word) -> usize {
-            let p_num = p as usize;
-            let words = p_num >> 3;
-            let pages = words >> 18;
-            pages.wrapping_mul(words)
-        }
-        let s = &*Slag::find(item, meta.total_bytes);
-        let rc_ptr = &s.rc as *const _ as *mut RefCount;
-        let (word, word_ix) = Slag::get_word(s.as_raw(), item, meta);
-        let word_ptr = ((s.as_raw() as *mut u8).offset(meta.bitset_offset) as *mut Word)
-            .offset(word);
-        let bucket_ind = self.bucket_num(hash_ptr(word_ptr));
-        let bucket = &mut *self.0.get(bucket_ind);
-        // XXX: using the property of the creek implementation that it fills newly-dirtied pages
-        // with zeros.
-        if bucket.rc.is_null() {
-            *bucket = RemoteFreeCell {
-                rc: rc_ptr,
-                word: word_ptr,
-                mask: 1 << word_ix,
-            };
-            // append this to a stack of buckets we know are non-null
-            self.1.push(bucket as *const _ as *mut u8);
-            return true;
-        }
-        if bucket.word == word_ptr {
-            bucket.mask |= 1 << word_ix;
-            return true;
-        }
-        false
-    }
-}
-
-/// A different approach to caching to `LocalCache` inspired by Bonwick-style magazines.
-///
-/// The advantage `MagazineCache` has over `LocalCache` is that it will unconditionally push to a
-/// local data-structure if there is room (it doesn't have to be a local free). The magazine
-/// structure also allows us to batch together remote frees (by simply composing bit-set masks
-/// ahead of a fetch-or), reducing the number of atomic instruction that must be issued for most
-/// remote frees.
-pub struct MagazineCache<CA: CoarseAllocator> {
-    stack_size: usize,
-    s: PtrStack,
-    iter: AllocIter,
-    alloc: SlagAllocator<CA>,
-    coalescer: Coalescer,
-}
-
-impl<CA: CoarseAllocator> LazyInitializable for MagazineCache<CA> {
-    type Params = (*mut Metadata, usize, CA, RevocablePipe<Slag>);
-    fn init(&(meta, decommit, ref page_alloc, ref avail): &Self::Params) -> Self {
-        let salloc = SlagAllocator::partial_new(meta, decommit, page_alloc.clone(), avail.clone());
-        Self::new(salloc)
-    }
-}
-
-impl<CA: CoarseAllocator> LazyInitializable for LocalCache<CA> {
-    type Params = (*mut Metadata, usize, CA, RevocablePipe<Slag>);
-    fn init(&(meta, decommit, ref page_alloc, ref avail): &Self::Params) -> Self {
-        let salloc = SlagAllocator::partial_new(meta, decommit, page_alloc.clone(), avail.clone());
-        Self::new(salloc)
-    }
-}
-
-impl<CA: CoarseAllocator> Drop for MagazineCache<CA> {
-    fn drop(&mut self) {
-        unsafe {
-            let meta = &*self.alloc.m;
-            let mask = self.iter.cur_word;
-            let word = self.iter.next_word.offset(-1);
-            let slag = self.alloc.slag;
-            // bulk-free the current AllocIter word. Then free all elements in the magazine.
-            self.alloc.bulk_free(mask, word, slag, meta);
-            for i in 0..self.s.top {
-                let item = *self.s.data.get(i);
-                self.alloc.free(item)
-            }
-        }
-    }
-}
-
-impl<CA: CoarseAllocator> Clone for MagazineCache<CA> {
-    fn clone(&self) -> Self {
-        MagazineCache::new_sized(self.alloc.clone(), self.stack_size)
-    }
-}
-
-impl<CA: CoarseAllocator> MagazineCache<CA> {
-    pub fn new_sized(mut alloc: SlagAllocator<CA>, magazine_size: usize) -> Self {
-        assert!(magazine_size > 0);
-        let s = PtrStack::new(magazine_size);
-        let iter = unsafe { alloc.refresh() };
-        let buckets = Coalescer::new(magazine_size * 2);
-        MagazineCache {
-            stack_size: magazine_size,
-            s: s,
-            iter: iter,
-            alloc: alloc,
-            coalescer: buckets,
-        }
-    }
-
-    pub fn new(alloc: SlagAllocator<CA>) -> Self {
-        use std::cmp;
-        let object_size = unsafe { (*alloc.m).object_size };
-        const CUTOFF: usize = 32 << 10;
-        let magazine_size = match object_size {
-            0 ... 512 => 1 << 16,
-            513 ... CUTOFF => 512 << 10 / object_size,
-            _ => 1 << 20 / object_size,
-        };
-        Self::new_sized(alloc, cmp::max(1, magazine_size))
-    }
-
-    /// Allocate memory from the current owned `Slag`.
-    ///
-    /// This amounts to getting memory from the current alloc iterator. If the iterator is
-    /// exhausted, a new `Slag` is acquired.
-    unsafe fn slag_alloc(&mut self) -> *mut u8 {
-        for _ in 0..2 {
-            match self.iter.next() {
-                Some(ptr) => return ptr,
-                None => self.iter = self.alloc.refresh(),
-            }
-        }
-        panic!(
-            "New slag is empty {:?} {:?}",
-            self.alloc.slag,
-            (*self.alloc.slag).rc.load()
-        )
-    }
-
-    pub unsafe fn alloc(&mut self) -> *mut u8 {
-        if let Some(ptr) = self.s.pop() {
-            trace_event!(cache_alloc);
-            ptr
-        } else {
-            trace_event!(slag_alloc);
-            self.slag_alloc()
-        }
-    }
-
-    pub unsafe fn free(&mut self, item: *mut u8) {
-        trace_event!(local_free);
-        if likely(self.s.top < self.stack_size) {
-            self.s.push(item);
-            return;
-        }
-        self.return_memory();
-        self.s.push(item);
-    }
-
-    /// Perform the bulk-level frees for the `Coalescer`.
-    unsafe fn return_memory(&mut self) {
-        debug_assert_eq!(self.s.top as usize, self.stack_size);
-        let new_top = self.stack_size / 2;
-        let meta = &*self.alloc.m;
-        // iterate over the stack and attempt to add them to the coalescer.
-        for i in new_top..self.stack_size {
-            let item = *self.s.data.get(i);
-            if !self.coalescer.insert(item, meta) {
-                // there was a "hash collision", so we simply free `item` directly
-                self.alloc.free(item)
-            }
-        }
-        self.s.top = new_top;
-        for cell_ptr in 0..self.coalescer.1.top {
-            let cell = &mut **(self.coalescer.1.data.get(cell_ptr) as *mut *mut RemoteFreeCell);
-            // Slag::find will technically work if you hand it any pointer within the slag
-            // itself, not just an object. As a result, we use the reference count to get at
-            // the slag it belongs to.
-            let slag = Slag::find(cell.rc as *mut u8, meta.total_bytes);
-            self.alloc.bulk_free(cell.mask, cell.word, slag, meta);
-            ptr::write(cell, RemoteFreeCell::default());
-        }
-        self.coalescer.1.top = 0;
-    }
-}
-
-/// A `LocalCache` provides thread-local data on top of a `SlagAllocator`.
-///
-/// Like a `MagazineCache`, it includes a stack of pointers to cache allocations. Unlike
-/// `MagazineCache`, only pointers local to the current `Slag` are placed in this stack: the rest
-/// are eagerly freed back to their owning `Slag`. Generally speaking, `LocalCache`s are better at
-/// producer-consumer workloads and worse at predominantly thread-local workloads than the
-/// `MagazineCache`. `LocalCache`s also have stronger guarantees when it comes to object locality:
-/// almost all adjacent allocations will come from the same region of memory.
-pub struct LocalCache<CA: CoarseAllocator> {
-    alloc: SlagAllocator<CA>,
-    vals: PtrStack,
-    iter: AllocIter,
-}
-
-impl<CA: CoarseAllocator> Drop for LocalCache<CA> {
-    fn drop(&mut self) {
-        unsafe {
-            let meta = &*self.alloc.m;
-            let mask = self.iter.cur_word;
-            let word = self.iter.next_word.offset(-1);
-            let slag = self.alloc.slag;
-            self.alloc.bulk_free(mask, word, slag, meta);
-            for i in 0..self.vals.top {
-                let item = *self.vals.data.get(i);
-                self.alloc.free(item)
-            }
-        }
-    }
-}
-impl<CA: CoarseAllocator> Clone for LocalCache<CA> {
-    fn clone(&self) -> LocalCache<CA> {
-        LocalCache::new(self.alloc.clone())
-    }
-}
-
-impl<CA: CoarseAllocator> LocalCache<CA> {
-    fn new(mut alloc: SlagAllocator<CA>) -> Self {
-        unsafe {
-            let stack = PtrStack::new((*alloc.m).n_objects);
-            let iter = alloc.refresh();
-            LocalCache {
-                alloc: alloc,
-                vals: stack,
-                iter: iter,
-            }
-        }
-    }
-
-    pub unsafe fn free(&mut self, it: *mut u8) {
-        if self.alloc.contains(it) {
-            self.vals.push(it);
-        } else {
-            self.alloc.free(it);
-        }
-    }
-
-    pub unsafe fn alloc(&mut self) -> *mut u8 {
-        self.vals
-            .pop()
-            .or_else(|| self.iter.next())
-            .unwrap_or_else(|| {
-                let next_iter = self.alloc.refresh();
-                self.iter = next_iter;
-                self.iter.next().expect("New iterator should have values")
-            })
-    }
-}
 /// A `DirtyFn` is a callback that is called upon allocating a clean page from a `PageAlloc`. It
 /// generally does nothing, but its presence in `PageAlloc` allows us to inject other callbacks for
 /// debugging or performance analysis.
@@ -1284,176 +1026,14 @@ impl<C: MemorySource, D: DirtyFn> CoarseAllocator for PageAlloc<C, D> {
     }
 }
 
-/// A thread-local stack data-structure for caching allocations from an owned `Slag`.
-///
-/// This implementation is specialized in a few ways:
-///
-/// * While `mem` has a maximum size, it is assumed that all `push` calls will leave the
-///   stack in a valid state. This is because the stack is initialized to have sufficient space
-///   for an entire `Slag` of pointers for this particular size class, and all `push`es only
-///   happen after the item in question has been confirmed to belong to a particular `Slag`.
-///   Additional checking must be implemented on top of `PtrStack` if these are not the desired
-///   semantics. See the `MagazineCache` implementation for an example of this.
-///
-/// * The stack is backed by an `OwnedArray`, which manages its own memory through `mmap`. This has
-///   a number of advantages.  It reduces reliance on `Vec`-like structures that are tied to the
-///   underlying `malloc` implementation. It also gives us lazy initialization without any extra
-///   work. Fresh  will be uncommited: this means that potentially large allocations of memory for
-///   stacks will only consume physical space when they are used.
-struct PtrStack {
-    data: OwnedArray<*mut u8>,
-    top: usize,
-}
-
-impl PtrStack {
-    fn new(max_objects: usize) -> PtrStack {
-        PtrStack {
-            data: OwnedArray::new(max_objects),
-            top: 0,
-        }
-    }
-
-    unsafe fn push(&mut self, item: *mut u8) {
-        *self.data.get(self.top) = item;
-        self.top += 1;
-    }
-
-    unsafe fn pop(&mut self) -> Option<*mut u8> {
-        if self.empty() {
-            None
-        } else {
-            self.top -= 1;
-            Some(*self.data.get(self.top))
-        }
-    }
-
-    #[inline]
-    fn empty(&self) -> bool {
-        self.top == 0
-    }
-}
-
-/// A builder-pattern-style builder for `MagazineAllocator`s and `LocalAllocator`s.
-///
-/// ```rust,ignore
-/// // A usize-specific allocator using the `LocalCache` frontend, customized to use 32K pages.
-/// let la: LocalAllocator<usize> = AllocBuilder::default().page_size(32 << 10).build_local();
-/// ```
-///
-/// Modifying other builder parameters past the default is not recommended. The overall API is
-/// unstable.
-pub struct AllocBuilder<T> {
-    cutoff_factor: f64,
-    page_size: usize,
-    target_overhead: usize,
-    eager_decommit_threshold: usize,
-    max_objects: usize,
-    _marker: PhantomData<T>,
-}
-
-impl<T> Default for AllocBuilder<T> {
-    fn default() -> Self {
-        AllocBuilder {
-            cutoff_factor: 0.6,
-            page_size: cmp::max(32 << 10, mem::size_of::<T>() * 4),
-            target_overhead: 1 << 20,
-            eager_decommit_threshold: 128 << 10,
-            max_objects: 1 << 30,
-            _marker: PhantomData,
-        }
-    }
-}
-
-impl<T> AllocBuilder<T> {
-    pub fn cutoff_factor(&mut self, cutoff_factor: f64) -> &mut Self {
-        self.cutoff_factor = cutoff_factor;
-        self
-    }
-    pub fn page_size(&mut self, page_size: usize) -> &mut Self {
-        self.page_size = page_size;
-        self
-    }
-    pub fn target_overhead(&mut self, target_overhead: usize) -> &mut Self {
-        self.target_overhead = target_overhead;
-        self
-    }
-    pub fn eager_decommit_threshold(&mut self, eager_decommit_threshold: usize) -> &mut Self {
-        self.eager_decommit_threshold = eager_decommit_threshold;
-        self
-    }
-    pub fn max_objects(&mut self, max_objects: usize) -> &mut Self {
-        self.max_objects = max_objects;
-        self
-    }
-
-    /// Build a `LocalAllocator<T>` from the current configuration.
-    pub fn build_local(&self) -> LocalAllocator<T> {
-        LocalAllocator::new_standalone(
-            self.cutoff_factor,
-            self.page_size,
-            self.target_overhead,
-            self.eager_decommit_threshold,
-            self.max_objects,
-        )
-    }
-
-    /// Build a `MagazineAllocator<T>` from the current configuration.
-    pub fn build_magazine(&self) -> MagazineAllocator<T> {
-        MagazineAllocator::new_standalone(
-            self.cutoff_factor,
-            self.page_size,
-            self.target_overhead,
-            self.eager_decommit_threshold,
-            self.max_objects,
-        )
-    }
-}
-
-macro_rules! typed_wrapper {
-    ($name:ident, $wrapped:tt) => {
-        pub struct $name<T>($wrapped<PageAlloc<MmapSource>>, PhantomData<T>);
-        impl<T> Clone for $name<T> {
-            fn clone(&self) -> Self {
-                $name(self.0.clone(), PhantomData)
-            }
-        }
-
-        impl<T> $name<T> {
-            pub fn new_standalone(cutoff_factor: f64,
-                                  page_size: usize,
-                                  target_overhead: usize,
-                                  eager_decommit: usize,
-                                  max_objects: usize)
-                -> Self {
-                    let pa = PageAlloc::new(page_size, target_overhead, 8, AllocType::SmallSlag);
-                    let slag = SlagAllocator::new(max_objects, mem::size_of::<T>(), 0,
-                                                  cutoff_factor, eager_decommit, pa);
-                    $name($wrapped::new(slag), PhantomData)
-                }
-
-            pub unsafe fn alloc(&mut self) -> *mut T {
-                self.0.alloc() as *mut T
-            }
-
-            pub unsafe fn free(&mut self, item: *mut T) {
-                self.0.free(item as *mut u8)
-            }
-        }
-        unsafe impl<T> Send for $name<T> {}
-    };
-}
-
-typed_wrapper!(LocalAllocator, LocalCache);
-typed_wrapper!(MagazineAllocator, MagazineCache);
-
 /// Allocator state wrapping a `Slag`.
 ///
 /// This struct forms the "backend" for a particular thread-local cache. It handles the state
 /// transitions of different `Slag`s and also acquires new `Slag`s for iteration over the bitset.
 pub struct SlagAllocator<CA: CoarseAllocator> {
-    m: *mut Metadata,
+    pub m: *mut Metadata,
     /// The current (local) `Slag`.
-    slag: *mut Slag,
+    pub slag: *mut Slag,
     /// Global pages, potentially not initialized to match `m`
     pages: CA,
     /// Available `Slag`s with metadata matching `m`.
@@ -1539,7 +1119,7 @@ impl<CA: CoarseAllocator> SlagAllocator<CA> {
     }
 
     /// Re-initialize a non-empty `AllocIter`; potentially getting a new `Slag` to do so.
-    unsafe fn refresh(&mut self) -> AllocIter {
+    pub unsafe fn refresh(&mut self) -> AllocIter {
         let s_ref = &*self.slag;
         let meta = &*self.m;
         let (_claimed, was) = s_ref.rc.unclaim();
@@ -1624,7 +1204,13 @@ impl<CA: CoarseAllocator> SlagAllocator<CA> {
         // safely return without further work.
     }
 
-    unsafe fn bulk_free(&mut self, mask: usize, word: *mut Word, slag: *mut Slag, meta: &Metadata) {
+    pub unsafe fn bulk_free(
+        &mut self,
+        mask: usize,
+        word: *mut Word,
+        slag: *mut Slag,
+        meta: &Metadata,
+    ) {
         let n_ones = mask.count_ones() as usize;
         if n_ones == 0 {
             return;
@@ -1652,7 +1238,7 @@ impl<CA: CoarseAllocator> SlagAllocator<CA> {
     }
 
     /// Perform a "remote" free to the `Slag` containing `item`.
-    unsafe fn free(&mut self, item: *mut u8) {
+    pub unsafe fn free(&mut self, item: *mut u8) {
         trace_event!(remote_free);
         let meta = &*self.m;
         let it_slag = Slag::find(item, meta.total_bytes);
@@ -1664,7 +1250,7 @@ impl<CA: CoarseAllocator> SlagAllocator<CA> {
     }
 
     /// Test if `it` is an element of the current `Slag`.
-    fn contains(&self, it: *mut u8) -> bool {
+    pub fn contains(&self, it: *mut u8) -> bool {
         unsafe {
             let meta = self.m.as_ref().expect("[contains] null metadata");
             let it_slag = Slag::find(it, meta.total_bytes);
@@ -1678,7 +1264,12 @@ impl<CA: CoarseAllocator> Clone for SlagAllocator<CA> {
         let mut new_page_handle = self.pages.clone();
         let first_slag = unsafe { new_page_handle.alloc() as *mut Slag };
         unsafe {
-            Slag::init(first_slag, self.m.as_ref().expect("[SlagAllocator::clone] null metadata"));
+            Slag::init(
+                first_slag,
+                self.m.as_ref().expect(
+                    "[SlagAllocator::clone] null metadata",
+                ),
+            );
         };
         SlagAllocator {
             m: self.m,
@@ -1688,157 +1279,4 @@ impl<CA: CoarseAllocator> Clone for SlagAllocator<CA> {
             eager_decommit_threshold: self.eager_decommit_threshold,
         }
     }
-}
-
-
-
-#[cfg(test)]
-mod tests {
-    extern crate env_logger;
-    use super::*;
-    use std::thread;
-    use std::ptr::write_volatile;
-    use std::collections::HashSet;
-
-    #[test]
-    fn metadata_basic() {
-        let _ = env_logger::init();
-
-        compute_metadata(8, 4096, 0, 0.8, 4, AllocType::SmallSlag);
-        compute_metadata(16, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
-        compute_metadata(24, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
-        compute_metadata(127, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
-        compute_metadata(800, 2 << 20, 0, 0.8, 32 << 10, AllocType::SmallSlag);
-        compute_metadata(514, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
-        compute_metadata(513, 2 << 20, 0, 0.8, 1024, AllocType::SmallSlag);
-        compute_metadata(768, 4096, 0, 0.8, 1024, AllocType::SmallSlag);
-        compute_metadata(800, 4096, 0, 0.8, 32 << 10, AllocType::SmallSlag);
-        compute_metadata(1025, 4096, 0, 0.8, 32 << 10, AllocType::SmallSlag);
-
-    }
-
-    #[test]
-    fn obj_alloc_basic() {
-        let _ = env_logger::init();
-        let mut oa = AllocBuilder::<usize>::default()
-            .page_size(4096)
-            .build_local();
-        unsafe {
-            let item = oa.alloc();
-            write_volatile(item, 10);
-            oa.free(item);
-        }
-    }
-
-    #[test]
-    fn obj_alloc_many_pages_single_threaded_usize() {
-        obj_alloc_many_pages_single_threaded::<usize>();
-    }
-
-    #[test]
-    fn obj_alloc_many_pages_single_threaded_u24() {
-        obj_alloc_many_pages_single_threaded::<[u8; 24]>();
-    }
-
-    #[test]
-    fn obj_alloc_many_pages_single_threaded_u32() {
-        obj_alloc_many_pages_single_threaded::<[u8; 32]>();
-    }
-
-    fn obj_alloc_many_pages_single_threaded<T: 'static>() {
-        let _ = env_logger::init();
-        const N_ITEMS: usize = 4096 * 20;
-        let mut oa = AllocBuilder::<T>::default().page_size(4096).build_local();
-        assert!(mem::size_of::<T>() >= mem::size_of::<usize>());
-        // stay in a local cache
-        for _ in 0..N_ITEMS {
-            unsafe {
-                let item = oa.alloc();
-                write_volatile(item as *mut usize, 10);
-                oa.free(item);
-            }
-        }
-
-        let mut v = Vec::with_capacity(N_ITEMS);
-        let mut h = HashSet::new();
-        for i in 0..N_ITEMS {
-            unsafe {
-                let item = oa.alloc();
-                write_volatile(item as *mut usize, i + 1);
-                v.push(item);
-                let item_num = item as usize;
-                assert!(!h.contains(&item_num));
-                h.insert(item_num);
-            }
-        }
-
-        for i in v {
-            unsafe {
-                oa.free(i);
-            }
-        }
-    }
-    #[test]
-    fn obj_alloc_many_pages_many_threads_usize() {
-        obj_alloc_many_pages_many_threads::<usize>()
-    }
-
-    #[test]
-    fn obj_alloc_many_pages_many_threads_u24() {
-        obj_alloc_many_pages_many_threads::<[u8; 24]>()
-    }
-
-    #[test]
-    fn obj_alloc_many_pages_many_threads_u32() {
-        obj_alloc_many_pages_many_threads::<[u8; 32]>()
-    }
-
-    fn obj_alloc_many_pages_many_threads<T: 'static>() {
-        let _ = env_logger::init();
-        use std::mem;
-        const N_ITEMS: usize = 4096 * 4;
-        const N_THREADS: usize = 40;
-        // TODO make macros for these tests and test both MagazineAllocator and LocalAllocator
-        let oa = AllocBuilder::<T>::default()
-            .page_size(4096)
-            .build_magazine();
-        // stay in a local cache
-        assert!(mem::size_of::<T>() >= mem::size_of::<usize>());
-        let mut threads = Vec::new();
-        for _ in 0..N_THREADS {
-            let mut my_alloc = oa.clone();
-            threads.push(thread::spawn(move || {
-                for _ in 0..N_ITEMS {
-                    unsafe {
-                        let item = my_alloc.alloc();
-                        write_volatile(item as *mut usize, 10);
-                        my_alloc.free(item);
-                    }
-                }
-
-                let mut v = Vec::with_capacity(N_ITEMS);
-                let mut h = HashSet::new();
-                for i in 0..N_ITEMS {
-                    unsafe {
-                        let item = my_alloc.alloc();
-                        write_volatile(item as *mut usize, i);
-                        v.push(item);
-                        let item_num = item as usize;
-                        assert!(!h.contains(&item_num));
-                        h.insert(item_num);
-                    }
-                }
-
-                for i in v {
-                    unsafe {
-                        my_alloc.free(i);
-                    }
-                }
-            }));
-        }
-        for t in threads {
-            t.join().expect("threads should exit successfully");
-        }
-    }
-
 }
